frontend/jupyter: avoid next build error by splittig up a file
