+++ src/packages/frontend/components/run-button/kernel-info.ts
@@ -3,18 +3,9 @@
  *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
  */
 
-import { fromJS } from "immutable";
 import LRU from "lru-cache";
-import { useEffect, useMemo, useState } from "react";
-import useAsyncEffect from "use-async-effect";
 
-import { useProjectContext } from "@cocalc/frontend/project/context";
 import type { KernelSpec } from "@cocalc/jupyter/types";
-import {
-  Kernels,
-  get_kernel_selection,
-  get_kernels_by_name_or_language,
-} from "@cocalc/jupyter/util/misc";
 import { capitalize } from "@cocalc/util/misc";
 import api from "./api";
 
@@ -94,56 +85,4 @@ export function kernelLanguage(
     }
   }
   return name;
-}
-
-// This returns processed jupyter kernel specs
-// ATTN: this is only valid inside the context of a specific project!
-export function useJupyterKernelsInfo() {
-  const [cnt, setCnt] = useState<number>(0);
-  const [error, setError] = useState<string>("");
-  const [kernelSpecs, setKernelSpecs] = useState<Kernels | null>();
-  const { isRunning, project_id } = useProjectContext();
-
-  function refresh() {
-    setCnt((cnt) => cnt + 1);
-  }
-
-  useEffect(() => refresh(), [isRunning]);
-
-  useAsyncEffect(async () => {
-    try {
-      const kernelInfo = await getKernelInfo(project_id, isRunning);
-      setKernelSpecs(fromJS(kernelInfo) as Kernels);
-    } catch (err) {
-      setError(`${err}`);
-    }
-  }, [project_id, cnt]);
-
-  const { kernel_selection, kernels_by_name, kernels_by_language } =
-    useMemo(() => {
-      if (kernelSpecs != null) {
-        const kernel_selection = get_kernel_selection(kernelSpecs);
-        const [kernels_by_name, kernels_by_language] =
-          get_kernels_by_name_or_language(kernelSpecs);
-        return {
-          kernel_selection,
-          kernels_by_name,
-          kernels_by_language,
-        };
-      } else {
-        return {
-          kernel_selection: null,
-          kernels_by_name: null,
-          kernels_by_language: null,
-        };
-      }
-    }, [kernelSpecs, error, project_id]);
-
-  return {
-    kernel_selection,
-    kernels_by_name,
-    kernels_by_language,
-    refresh,
-    error,
-  };
 }
